Centralize all notification fetching and menus
